Prototype for 2.5D Game - Initial commit.
Meant to provide a global networked game, portable or for arcade machines.
Mostly improvised and hardcoded stuff, just enough to make it work and have something to work and improve later.
It takes (hardcoded) map data in the same format as a doom wad would provide (segments, sectors, etc) and renders it in a similar way. It divides the drawing window in vertical slices and draws each wall, rool and ceiling as a collection of columns. It tries to avoid overdraw as much as possible, but the BSP and culling algorithms optimizations are not implemented yet. Textures can work but I disabled them for now (Still need to fix some scaling and UV calculations)
Still needs fixes for sprite and window scaling, test networking and improve collisions.
All this should be refactored too in the future, so i can provide a common interface to build games easily, like a mini game engine, add better keyboard handling, menus and such.